backend/python-server/requirements.txt
python-dotenv==1.0.1
pydantic-settings>=2.0.0

langchain-openai>=0.3.0
langchain-core>=0.3.0

fastapi>=0.115.0
uvicorn[standard]==0.32.0

psycopg2-binary
sqlalchemy>=2.0.0

redis>=5.0.0